Refactor views and models to be in their own directories
