+++ lib/home.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:file_picker/file_picker.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:flutter/services.dart';
-import 'task_list_view.dart';
+import 'views/task_list_view.dart';
 import 'settings_view.dart';
 import 'task_create_dialog.dart';
