+++ lib/models/upcoming_list_item.dart
@@ -1,4 +1,4 @@
-import '../task_model.dart';
+import '../models/task_model.dart';
 
 class UpcomingListItem {
   final bool isHeader;
